Report versions by year (#135)
