+++ package.json
@@ -54,7 +54,7 @@
     "serverless-offline": "^14.4.0",
     "serverless-plugin-scripts": "^1.0.2",
     "serverless-stack-termination-protection": "^2.0.2",
-    "typescript": "^4.6.3",
+    "typescript": "^5.8.2",
     "util": "^0.12.4",
     "yargs": "^16.1.1"
   }
