+++ services/app-api/handlers/reports/buildReport.test.ts
@@ -26,7 +26,9 @@ describe("Test create report handler", () => {
     const reportOptions = {
       name: "report1",
       year: 2026,
-      options: {},
+      options: {
+        pom: true,
+      },
     } as unknown as ReportOptions;
     const report = await buildReport(
       ReportType.QMS,
