+++ services/app-api/handlers/reports/buildReport.ts
@@ -1,22 +1,22 @@
 import KSUID from "ksuid";
-import { qmsReportTemplate } from "../../forms/qms";
+import {
+  getReportTemplate,
+  getCmitInfo,
+} from "../../forms/yearlyFormSelection";
 import { putReport } from "../../storage/reports";
 import {
-  ElementType,
-  PageElement,
   Report,
   ReportStatus,
   ReportOptions,
   ReportType,
+  isHeaderTemplate,
+  MeasureOptions,
+  MeasurePageTemplate,
+  CMIT,
 } from "../../types/reports";
 import { User } from "../../types/types";
-import { CMIT_LIST } from "../../forms/cmit";
 import { validateReportPayload } from "../../utils/reportValidation";
 import { logger } from "../../libs/debug-lib";
-
-const reportTemplates = {
-  [ReportType.QMS]: qmsReportTemplate,
-};
 
 export const buildReport = async (
   reportType: ReportType,
@@ -24,7 +24,9 @@ export const buildReport = async (
   reportOptions: ReportOptions,
   user: User
 ) => {
-  const report = structuredClone(reportTemplates[reportType]) as Report;
+  const year = reportOptions.year;
+  const report = structuredClone(getReportTemplate(reportType, year));
+  const cmitList = getCmitInfo(year);
   report.state = state;
   report.id = KSUID.randomSync().string;
   report.created = Date.now();
@@ -49,26 +51,17 @@ export const buildReport = async (
       measures.push(...report.measureLookup.pomMeasures);
     }
 
-    const measurePages = measures.map((measure) => {
-      const pages = measure.measureTemplate.map((template) =>
-        structuredClone(report.measureTemplates[template])
+    for (let measure of measures) {
+      const cmitInfo = cmitList.find((cmit) => cmit.uid === measure.uid)!;
+      const pages = measure.measureTemplate.map((templateName) =>
+        initializeQmsPage(
+          measure,
+          report.measureTemplates[templateName],
+          cmitInfo
+        )
       );
-
-      return pages.map((page) => {
-        page.cmit = measure.cmit;
-        page.cmitId = measure.uid;
-        page.stratified = measure.stratified;
-        page.required = measure.required;
-        page.elements = [
-          ...page.elements.map((element) =>
-            findAndReplace(element, measure.uid)
-          ),
-        ];
-        return page;
-      });
-    });
-
-    report.pages = report.pages.concat(...measurePages);
+      report.pages.push(...pages);
+    }
   }
 
   /**
@@ -88,12 +81,30 @@ export const buildReport = async (
   return report;
 };
 
-export const findAndReplace = (element: PageElement, uid: string) => {
-  const cmitInfo = CMIT_LIST.find((list) => list.uid === uid);
-  if (cmitInfo) {
-    if (element.type === ElementType.Header) {
-      element.text = element.text.replace("{measureName}", cmitInfo.name);
+/**
+ * Clone the given template, and fill it in with the necessary data.
+ */
+const initializeQmsPage = (
+  measure: MeasureOptions,
+  template: MeasurePageTemplate,
+  cmitInfo: CMIT
+) => {
+  const page = structuredClone(template);
+  page.cmit = measure.cmit;
+  page.cmitId = measure.uid;
+  page.stratified = measure.stratified;
+  page.required = measure.required;
+  for (let i = 0; i < page.elements.length; i += 1) {
+    let element = page.elements[i];
+    if (isHeaderTemplate(element)) {
+      /*
+       * Many pages share the same `measureHeader` object, from elements.ts
+       * The extra clone ensures we only alter this page's header.
+       */
+      const clone = structuredClone(element);
+      clone.text = clone.text.replace("{measureName}", cmitInfo.name);
+      page.elements[i] = clone;
     }
   }
-  return element;
+  return page;
 };
